@@ -1,8 +1,6 @@
 import { getLocation, dataQaExists } from './utils'
 import { HOST } from './config'
-
 import { fillAnonymousLogin, acceptBet } from './helpers'
-
 import * as api from './api'
 
 fixture`Anonymous login by accepting`.page`${HOST}`
@@ -17,7 +15,7 @@ test('I can accept a bet by loggin in', async t => {
 
   await t
     .expect(await getLocation())
-    .eql(`/anonymous-login`, 'redirects to anonymous login')
+    .eql(`/login`, 'redirects to anonymous login')
 
   await fillAnonymousLogin(t, `Creator's friend`)
 
