@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { withKnobs } from '@storybook/addon-knobs'
-import Box from './../default-container'
+import Box from '../default-container'
 
 import BetInput from '.'
 
